Refactor translatable Nova resources some
