+++ app/Nova/TranslatableResource.php
@@ -1,0 +1,76 @@
+<?php
+
+namespace App\Nova;
+
+use Laravel\Nova\Http\Requests\NovaRequest;
+use Laravel\Nova\Resource as NovaResource;
+
+abstract class TranslatableResource extends NovaResource
+{
+    /**
+     * Build an "index" query for the given resource.
+     *
+     * @param  \Laravel\Nova\Http\Requests\NovaRequest  $request
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+
+     public static function indexQuery(NovaRequest $request, $query)
+     {
+         return $query->orderByTranslation('name');
+     }
+
+     
+    /**
+     * Overrides to make search/filter work with translations.
+     */
+    protected static function applySearch($query, $search)
+    {
+        return $query->where(function ($query) use ($search) {
+            $model = $query->getModel();
+
+            foreach (static::searchableColumns() as $column) {
+                $query->orWhere($model->qualifyColumn($column), 'ilike', '%'.$search.'%');
+            }
+            $query->orWhereTranslationLike('name', '%'.$search.'%');
+        });
+    }
+
+    /**
+     * Build a Scout search query for the given resource.
+     *
+     * @param  \Laravel\Nova\Http\Requests\NovaRequest  $request
+     * @param  \Laravel\Scout\Builder  $query
+     * @return \Laravel\Scout\Builder
+     */
+    public static function scoutQuery(NovaRequest $request, $query)
+    {
+        return $query;
+    }
+
+    /**
+     * Build a "detail" query for the given resource.
+     *
+     * @param  \Laravel\Nova\Http\Requests\NovaRequest  $request
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public static function detailQuery(NovaRequest $request, $query)
+    {
+        return parent::detailQuery($request, $query);
+    }
+
+    /**
+     * Build a "relatable" query for the given resource.
+     *
+     * This query determines which instances of the model may be attached to other resources.
+     *
+     * @param  \Laravel\Nova\Http\Requests\NovaRequest  $request
+     * @param  \Illuminate\Database\Eloquent\Builder  $query
+     * @return \Illuminate\Database\Eloquent\Builder
+     */
+    public static function relatableQuery(NovaRequest $request, $query)
+    {
+        return parent::relatableQuery($request, $query);
+    }
+}
